failing test for root page title
for the initial pass we'll put the character list here. This will
probably be hidden behind auth later.

We're just asserting that the title is sane - for OWIN e2e tests I think
this is about as detailed as we're going to get. Already we have to
import HtmlAgility and that feels like overkill for what an e2e test
should look like